requirements.txt
# requirements.txt

asgiref==3.8.1
certifi==2024.8.30
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.6
